chore: adding utils.py
